June 13

+++ Polymorphism/src/method_Overriding/Triangle.java
@@ -4,7 +4,7 @@ public class Triangle extends Shape{
 	private double base;
 	private double height;
 	
-    public Triangle (double base,double height) {
+    public Triangle (int base,int height) {
     	this.base=base;
     	this.height=height;
     }
